Change to floats

+++ db/migrate/20210720083618_create_school_targets.rb
@@ -3,9 +3,9 @@ class CreateSchoolTargets < ActiveRecord::Migration[6.0]
     create_table :school_targets do |t|
       t.references :school, null: false, foreign_key: true
       t.date :target
-      t.integer :electricity
-      t.integer :gas
-      t.integer :storage_heaters
+      t.float :electricity
+      t.float :gas
+      t.float :storage_heaters
 
       t.timestamps
     end
